webpki: move WebPkiClientVerifier to client_verifier
This commit moves the `WebPkiClientVerifier` type (and associated bits)
from `verify.rs` into `client_verifier.rs` to live alongside its builder
type.
